Create Transform operation
Estimated time
15-20 minutes

Level of difficulty
Easy

Objectives
Familiarize the student with:

transform - using the function;
transform - using the function with a variable.
Scenario
Write a program that creates four vectors (as in the code below, we call these variables: valuesA, valuesB and valuesC, results).

Get one value from the user (we'll call it value), which will be added to our values. We have a simple equation where result = A + B * C + value, where A, B, and face="courier new">C are the values from the corresponding vectors.

Try to use as few transform operations as possible. All values are integer numbers. Print the results vector (values separated by a space) at the end of the program. Your version of the program must print the same result as the expected output for all given input. Test it for all input cases.


Example data:

Input:

1

Output:

6 9 8 13 8
output


Input:

0

Output:

5 8 7 12 7
output


Input:

5

Output:

10 13 12 17 12