Create Generating operation
Estimated time
15-20 minutes

Level of difficulty
Easy

Objectives
Familiarize the student with:

generate_n - using the function;
functional objects - writing them.
Scenario
Write a program that creates two vectors filled with triangular numbers. The first vector has a const size of six elements. The second vector size is not known at compilation time.

Ask the user how many of the first triangular numbers your program should generate into the second vector. A triangular number counts the objects that can form an equilateral triangle. The nth triangular number is the number of dots or balls in a triangle with n dots on each side. In fact, it's the sum of the n natural numbers from 1 to n. The first triangular number is 1, the next is 3 (1+2), the third is 6 (1+2+3) and the fourth is 10 (1+2+3+4).

Print all the triangular numbers from both vectors. Your version of the program must print the same result as the expected output for all given input. Test it for all input cases.


Example data:

Input:

5

Output:

1 3 6 10 15 21
1 3 6 10 15
output


Input:

6

Output:

1 3 6 10 15 21
1 3 6 10 15 21
output


Input:

11

Output:

1 3 6 10 15 21
1 3 6 10 15 21 28 36 45 55 66
output