Create Removing opearations
Estimated time
15-20 minutes

Level of difficulty
Easy

Objectives
Familiarize the student with:

generate - using the function;
remove - using the function;
erase - using the function;
printing on screen.
Scenario
Write a program that creates a vector and fills it with a sequence of values from 1 to a value given by the user. Next, your program should remove all values which are divisible by 3 (without the remainder) or have the digit 3 in them. After this, print only the remaining values. Don't forget that remove_if returns an iterator to the end of the remaining sequence.

Your version of the program must print the same result as the expected output for all given input. Test it for all input cases.


Example data:

Input:

10

Output:

1 2 4 5 7 8 10
output


Input:

19

Output:

1 2 4 5 7 8 10 11 14 16 17 19
output


Input:

39

Output:

1 2 4 5 7 8 10 11 14 16 17 19 20 22 25 26 28 29
output